requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
django-cors-headers==4.3.1
gunicorn==23.0.0
packaging==24.2
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.8.2
psycopg2-binary==2.9.10
reportlab==4.2.5
Pillow==11.0.0
django-storages==1.14.4
boto3==1.35.86
africastalking>=1.2.0
dj-database-url==2.1.0
